Isolates code into a library
