+++ server/content-routes.js
@@ -1,65 +1,18 @@
 const express = require("express");
-const { promisify } = require("util");
 const fs = require("fs");
 const path = require("path");
+const { promisify } = require("util");
+const { topicTitleIsUnique, addTopicToTree, saveAndSendContent, loadContent } = require("./lib");
 
 const readFile = promisify(fs.readFile);
-const writeFile = promisify(fs.writeFile);
 const router = express.Router();
-
-const topicTitlesOnly = topic => topic.title
-const nestedTitlesOnly = (subTitles, topic) => [ ...subTitles, ...titlesOnly(topic)];
-const titlesOnly = ({ agenda=[] }) => [...agenda.map(topicTitlesOnly), ...agenda.reduce(nestedTitlesOnly, [])];
-const matchCase = rightString => leftString => leftString.toLowerCase() === rightString.toLowerCase()
-const topicTitleIsUnique = (topicName, topic) => !titlesOnly(topic).some(matchCase(topicName));
-
-function addTopicToTree(content, { title, difficulty="beginner", type="section", length=1, required=true }) {
-  return {
-    ...content,
-    agenda: [
-      ...content.agenda,
-      {
-        title,
-        difficulty,
-        type,
-        length,
-        required,
-        agenda: []
-      }
-    ]
-  };
-}
-
-async function saveAndSendContent(res, content, rootFolder) {
-  try {
-    await writeFile(path.join(rootFolder, "timesplitter.json"), JSON.stringify(content, null, 2));
-  } catch (error) {
-    console.error(error);
-    res.json(error);
-  }
-  res.json(content);
-}
 
 module.exports = function (rootFolder) {
 
-  let content = fs.readFileSync(
-    path.join(rootFolder, "timesplitter.json"),
-    "UTF-8"
-  );
-  
-  try {
-    content = JSON.parse(content);
-  } catch (error) {
-    let clearError = new Error(
-      "There was an error parsing the timesplitter.json file"
-    );
-    console.error(clearError);
-    throw error;
-  }
+  let content = loadContent(rootFolder);
 
   router.post("/", async (req, res) => {
     const { topic, difficulty } = req.body;
-    console.log(`Adding ${topic} - ${difficulty}`)
     if (!topicTitleIsUnique(topic, content)) {
       const error = new Error(`topic title "${topic}" is not unique.`)
       console.error(error)
